Revert "add form integration test"
This reverts commit 97e439865bbe13ffa61f3d2d80cc4812009ab5f6.
